Time: 386 ms (9.09%), Space: 267.2 MB (9.09%) - LeetHub

+++ 3259-maximum-energy-boost-from-two-drinks/3259-maximum-energy-boost-from-two-drinks.cpp
@@ -2,11 +2,11 @@ class Solution {
 public:
     long long maxEnergyBoost(vector<int>& energyDrinkA, vector<int>& energyDrinkB) {
         int n=energyDrinkA.size();
-        vector<vector<int>> cache(n+1,vector<int>(2,-1));
+        vector<vector<long long>> cache(n+1,vector<long long>(2,-1));
         return max(solve(energyDrinkA,energyDrinkB,0,0,cache),solve(energyDrinkA,energyDrinkB,0,1,cache));
     }
 
-    long long solve(vector<int>& energyDrinkA, vector<int>& energyDrinkB,int i,int lastUsed,vector<vector<int>>& cache){
+    long long solve(vector<int>& energyDrinkA, vector<int>& energyDrinkB,int i,int lastUsed,vector<vector<long long>>& cache){
         if(i>=energyDrinkA.size()){
             return 0;
         }
